% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddIn_FindAndReplace.R
\name{AddIn_FindAndReplace}
\alias{AddIn_FindAndReplace}
\title{Find and Replace Tokens in a Document}
\usage{
AddIn_FindAndReplace()
}
\description{
Call this as an addin to replace occurrences of a particular sequence of
characters in a document with a new sequence of characters.
}
