% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{LoadPackages}
\alias{LoadPackages}
\title{Quietly load one or more packages}
\usage{
LoadPackages(libraries, stealth = FALSE)
}
\arguments{
\item{libraries}{a character vector of libraries to load}
}
\value{
None
}
\description{
Wraps install.packages and require/requireNamespace
}
\examples{
LoadPackages(c("stringr", "stringi"))

}
