% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_CleanDates}
\alias{str_CleanDates}
\title{#' @export
str_Trim <- function(.string){
  str_trim(.string, side = "both");
}}
\usage{
str_CleanDates(.string, .output_format = "\%Y-\%m-\%d",
  .input_format = c("Ymd", "mdY"))
}
\description{
#' @export
str_TrimLeft <- function(.string){
  str_trim(.string, side = "left");
}
}
\details{
#' @export
str_TrimRight <- function(.string){
  str_trim(.string, side = "right");
}
}
