% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_Detect}
\alias{str_Detect}
\title{#' @export
str_replace_all <- stringr::str_replace_all
#' @export
str_replace <- stringr::str_replace
#' @export
str_split_fixed <- stringr::str_split_fixed
#' @export
str_split <- stringr::str_split}
\usage{
str_Detect(string, pattern)
}
\description{
#' @export
str_replace_all <- stringr::str_replace_all
#' @export
str_replace <- stringr::str_replace
#' @export
str_split_fixed <- stringr::str_split_fixed
#' @export
str_split <- stringr::str_split
}
