% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_Trim}
\alias{str_Trim}
\title{#' @export
str_detect <- stringr::str_detect
#' @export
fixed <- stringr::fixed
#' @export
str_split <- stringr::str_split}
\usage{
str_Trim(...)
}
\description{
#' @export
str_detect <- stringr::str_detect
#' @export
fixed <- stringr::fixed
#' @export
str_split <- stringr::str_split
}
