% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{TwoLawsOptions}
\alias{TwoLawsOptions}
\title{Set or get options for my package}
\usage{
TwoLawsOptions(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{verbose}}{(\code{logical};FALSE) Whether or not to include verbose output useful for debugging }
}
}

